Simplify type management by removing types/ dir
