+++ redux/actions/parameters/add-new-line-in-parameter-array.ts
@@ -1,5 +1,7 @@
-// eslint-disable-next-line no-unused-vars
-import { AddNewLineInParameterArrayAction } from "../../../types/parameters";
+export interface AddNewLineInParameterArrayAction {
+  type: "ADD_NEW_LINE_IN_PARAMETER_ARRAY",
+  path: string;
+}
 
 export function addNewLineInParameterArray(path: string): AddNewLineInParameterArrayAction {
   return {
