+++ redux/actions/parameters/index.ts
@@ -1,5 +1,5 @@
-export { addNewLineInParameterArray } from "./add-new-line-in-parameter-array";
-export { removeLastLineInParameterArray } from "./remove-last-line-in-parameter-array";
-export { resetAmendementToBase } from "./reset-amendement-to-base";
-export { resetAmendementToPlf } from "./reset-amendement-to-plf";
-export { updateParameter } from "./update-parameter";
+export * from "./add-new-line-in-parameter-array";
+export * from "./remove-last-line-in-parameter-array";
+export * from "./reset-amendement-to-base";
+export * from "./reset-amendement-to-plf";
+export * from "./update-parameter";
