+++ redux/reducers/parameters/parameters-state.interface.ts
