--- types/parameters/index.ts
@@ -1,2 +1,0 @@
-export * from "./actions";
-export * from "./parameters-state";
